Refactored to use snapsToCsvDownload, snappablesToCsvDownload. Moved formatBody into snapFormatting.
