+++ app/src/components/submissionPage/helpers/snapFormatting.ts
@@ -33,3 +33,13 @@ export function getBodyElements(body: string): BodyElement[] {
     }
     return bodyElements;
 }
+
+/* Formats the body of a snap so the meta data is removed */
+export function formatBody(body: string) {
+    var newBody = "";
+    const bodyElems = getBodyElements(body);
+    for (var elem of bodyElems) {
+        newBody += elem.text;
+    }
+    return newBody;
+}
